requirements.txt
numpy
matplotlib
pytorch
scikit-learn
tqdm
seaborn
brewer2mpl
scipy